BMS-18 BUGFIX: Cell 1 & 2 were reading the same value.

+++ software/src/cell.h
@@ -30,7 +30,7 @@ enum cell_balancing {
 typedef enum cell_balancing cell_balancing_t;
 
 typedef struct {
-    uint8_t cell_id;
+    uint32_t cell_id;
     cell_status_t status;
     cell_health_t health;
     cell_balancing_t is_balancing;
